Fix dark and light mode icons
The dark mode and light mode icons were oerlapping
and showing at the same time in light mode. I
accidentally deleted the classes class="cs-moon"
and class="cs-sun" when switching out the icons.
Adding the classes back was the easiest solution
to fix the problem.

The sun icon is still broken. It needs to be white
so it shows up in the dark nav bar on dark mode.